pyproject.toml
[project]
name = "storitch"
version = "2.0"
description = "Simple file upload"
dependencies = [
    "fastapi==0.115.0",
    "pyyaml==6.0.2",
    "click==8.1.7",
    "aiofiles==24.1.0",
    "uvicorn==0.30.6",
    "python-multipart==0.0.10",
    "pydantic-settings==2.5.2",
    "pillow==10.4.0",
    "pydicom==3.0.1",
]

[project.optional-dependencies]
dev = ["requests==2.32.3", "httpx==0.27.2", "pytest==8.3.3"]


[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core>=3.2,<4"]

[tool.ruff]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]
